fix!: Change the format of a returned account locale to xx_YY
When the carbonio-user-management returns the locale (zimbraLocalPref)
of a requested account, the string value was returned in upper case
(XX_YY) and this raised a problem because the format of this value was
unreadable by other systems.

Now the carbonio-user-management returns the locale in the xx_YY format
checking if the locale returned by the carbonio-mailbox is a valid one.
This check is necessary since the value of this type of pref can be set
manually by a sysadmin without any check from the carbonio-mailbox.

When the system receives an invalid locale it falls back to 'en'.

BREAKING CHANGE: The /users/myself API response is changed to follow the
different format used to represent the locale value. Now it is a string
(instead of an enumerator) and it can support all the values defined in
the standard.

refs: UM-25